Types: add Base64UrlString
JWA §4.7.1. defines the "apu" (Agreement PartyUInfo) and "apv"
(Agreement PartyVInfo) ECDH Key Agreement header parameters, which
are defined as base64url encoded strings.

Add HSpec based on examples given in the JWA specification and
implement the Base64UrlString type and its FromJSON instance.  Note
that "=" padding may be absent from the input so it is necessary to
equals-pad the input to a multiple of four characters.