JWS: simplify JWS Header representation
In the existing JWS Header representation, "input data" is carried
explicitly in the data type, complicating the Eq instance.
Furthermore, on parse there is a loss of information about which
header fields are conveyed in the protected header or the
unprotected header.

Introduces the 'Armour' abstraction that encapsulates
partial decoding with remembering the input value.

Update the 'Signature' type to include two 'JWSHeader' values, one
armoured (the JWS Protected Header) and one bare (the JWS
Unprotected Header).

The additional expressiveness in having two 'JWSHeader' objects in
the 'Signature' comes (for now) with a loss of expressiveness in
enshrining certain restrictions of the standard in the type.  In
particular, the "alg" requirement and the requirement that no header
parameter appear in both the protected and the unprotected header
are no longer embodied in the types.  Pursuant to this, introduce
new error values that signal these violations of the spec and update
the parsing method to check for them.