Refactor Signature to have single JWSHeader
Currently Signature holds two (Maybe JWSHeader) values, each of
which contains a (Maybe Alg) field.  This is a poor model for the
JWS spec because:

- There must be at least a protected or unprotected header

- The JWS is conceptually a single value spread over the two
  protected and unprotected headers

- Duplicate header parameters are not allowed

- The "alg" parameter is required

Furthermore, we want a single header value so we can abstract that
out to allow for extension.

Add the HasParams class, Protection flag, HeaderParam type and
assorted helper functions for working with JOSE headers.

Change Signature to have exactly one JWSHeader.

Change JWSHeader to have exactly one Alg.

Change JWSHeader fields that may appear in either protected or
unprotected header to 'HeaderParam a' (which carries Protection).

Update JSON and Compact instances.