JWS: sign in unified MonadRandom / MonadError monad
Currently the type of signing operations is:

  (MonadRandom m) => m (Either Error r)

which can be generalised to:

  (MonadRandom m, MonadError e m1, AsError e) => m (m1 e r)

This does not compose well with operations which have the type:

  (MonadError e m1, AsError e) => m1 e r

But users will frequently need to compose these operations.  To make
them composable, change the type of functions that require both
randomness and error throwing to:

  (MonadRandom m, MonadError e m, AsError e) => m e r

'ExceptT' can be used to add errors to 'MonadRandom'.  A generic
'MonadRandom' instance for transformers is provided provided in
module 'Crypto.JOSE.Error'.