JWT: optics for StringOrURI
Provide optics for working with the StringOrURI type, and remove the
old getters and constructors.

The enforcement of the RFC 7519 requirement that value containing a
':' MUST be a URI now occurs in the prism, which is (directly or
indirectly) the only (exported) way to construct a StringOrURI.
Consequently, the IsString method is now non-total.

Fixes: https://github.com/frasertweedale/hs-jose/issues/34