JWS: parameterise over protection indicator
Add the 'ProtectionIndicator' class, and parameterise 'HeaderParam',
'JWSHeader', 'Signature' and 'JWS' over it.  Define instances for
'Protection' (protected or unprotected) and '()' (protected).

Update 'headerRequired' and 'headerOptional' parsers to fail when
they read an unprotected header, but the 'ProtectionIndicator' type
does not support unprotected headers.

Add the 'GeneralJWS', 'FlattenedJWS' and 'CompactJWS' type synonyms
as shorthand for various instantiations of 'JWS'.

Drive-by rename 'jWSHeader' to 'jwsHeader'.

Change type of 'encodeCompact' to always succeed, and restrict
'ToCompact' and 'FromCompact' instances to 'JWS Identity ()' so that
it is impossible to represent unprotected headers when working with
the compact serialisation.

Add tests to verify the decoding behaviour for 'Signature
Protection' (allows unprotected headers) and 'Signature ()'
(prohibits unprotected headers).