key size check: test RSA public number directly
The SizedBased64Integer decoding routine sets the size based on
number of octets, even if there are leading null bytes (which is
wrong, according to JWA spec).  When performing key size checks, we
check the recorded size.  If there are leading null bytes this can
lead to erroneous acceptance of a too-small key.

Update the RSA key size checks to ignore the recorded size and test
the public number directly.

Although SizedBase64Integer is not the correct type (it should be
Base64Integer in a number of places) the change of type will occur
in a separate commit so that this commit can be easily backported to
older branches.