JWK: hide ECKeyParameters constructor
Hide the ECKeyParameters constructor, prohibiting users from
twiddling the coordinates to make a point leave the curve.  This is
a countermeasure against invalid curve attacks.

The only places we construct ECKeyParameters are when generating a
key (we assume that cryptonite generates a point on the curve) and
when parsing (we check that the point is on the curve and fail if it
is not).