implement effectful JWKStore
A JWKStore might need to access a service or filesystem to look up
and retrieve keys.  A store might also want access to a JWS payload
to help it choose a key, e.g. using the "iss" claim in a signed JWT
to choose the public key for verification.

Implement effects in the JWKStore class members.  The instances for
JWK and JWKSet can be used in any Applicative.  Also add a type
variable to JWKStore for the payload type, so that a store can
inspect a payload.  This feature will be further developed in
subsequent commits.