support inspecting payload in JWKStore
JWKStore needs the ability to inspect a payload to help it choose
key(s).  For example, a JWT Claims Set "iss" value could identify
the verification key.

Add 'verifyJWSWithPayload' which, given a payload parser, passes the
payload to the JWKStore and returns the parsed payload if
verification succeeds.  This approach upholds the philosophy of not
providing any access to the payload except as the result of a
successful verification.

Update JWT verification to use this new payload decoding facility.
Now a JWT ClaimsSet-aware JWKStore can read the claims when choosing
a key.