rename JWKStore to VerificationKeyStore
With the addition of the payload parameter for key lookup, the
'keysFor' class member was no longer suited for encryption key
lookup.  Furthermore, expanding the class with a new function when
we add JWE support is not desirable (consider wanting different
classes for different JWS payload types).  Instead we will have
separate classes for verification key lookup and encryption key
lookup.

Therefore, rename 'JWKStore' and 'keysFor' to 'VerificationKeyStore'
and 'getVerificationKeys' respectively.  Also remove the 'keys'
member.