remove 'JWT' newtype
We did not export the JWT constructor, nor did we export an optic to
its inner part.  This meant that when working with a JWT, the JWS
header could not be read to find out the "kid" (or whatever).

In considering adding optics to this, it is observed that there is
no JWT-specific behaviour until the ClaimsSet gets parsed *during
verification*.  Furthermore, examples and presumably most users are
using the 'SignedJWT' type synonym for convenience, rather than 'JWT
(CompactJWS JWSHeader)'.

Simply removing the 'JWT' type is the best outcome.

Fixes: https://github.com/frasertweedale/hs-jose/issues/64