change x5c type to (NonEmpty X509.SignedCertificate)
There is no real reason to keep the newtype within the JWSHeader and
JWK data types themselves - it is only needed for parsing/printing
and is extraneous w.r.t. the data type definition.  Update the data
declarations and perform necessary changes to the FromJSON/ToJSON
instances.

Add 'instance Functor (HeaderParam p)' to support this change.

This change also uncovered a bug in JWSHeader parsing where the
X509-related fields were read from the wrong keys, which was fixed
in this commit.