reuse raw protected header when serialising
When deserialising the raw protected header is recorded. But
serialising ignores the raw protected header. If hs-jose (aeson)
serialises the protected header params in a different order from the
original signer, the resulting JWS is no longer valid.

Reuse the raw protected header (if available) when serialising a
Signature.

To make it possible to test this, export a new function:

  rawProtectedHeader
    :: (HasParams a, ProtectionIndicator p)
    => Signature p a -> B.ByteString

Also refactor some other functions to make use of this new function,
and perform some related cleanups.

Fixes: https://github.com/frasertweedale/hs-jose/issues/98