tests: convert (most) property tests to hedgehog
Converting to Hedgehog will allow getting rid of QuickCheck, which
is currently a dependency of both lib and tests.  Furthermore, there
are some orphan Arbitrary instances which are known to cause issues.

As a first step, convert most of the property tests.  Defer the JWK
round-trip test.  It requires a large number of Gen values, which
will be implemented in subsequent commit(s).