add JWT sign/verify variants general in payload type
Add variants of the JWT signing and verification functions that are
general in the payload type.  This provides for applications that
use additional claims beyond those registered by RFC 7519.

Update the tests to use an example extended payload type in addition
to ClaimsSet directly.

Examples in documentation still need to be updated.

The `unregisteredClaims` field in ClaimsSet is likely to be
deprecated in the next release, and removed in a future release.